+++ Cloudinary/Classes/Core/Features/Downloader/CLDDownloader.swift
@@ -38,24 +38,4 @@ import Foundation
     internal override init(networkCoordinator: CLDNetworkCoordinator) {
         super.init(networkCoordinator: networkCoordinator)
     }
-
-    // MARK: - Actions
-    /**
-    Asynchronously fetches a remote image from the specified URL.
-    The image is retrieved from the cache if it exists, otherwise its downloaded and cached.
-    
-    - parameter url:                    The image URL to download.
-    - parameter progress:          The closure that is called periodically during the data transfer.
-    - parameter completionHandler:      The closure to be called once the request has finished, holding either the retrieved UIImage or the error.
-    
-    - returns:              A `CLDFetchImageRequest` instance to be used to get the fetched image from, or to get the download progress or cancel the task.
-    */
-    @discardableResult
-    open func fetchImage(_ url: String, _ progress: ((Progress) -> Void)? = nil, completionHandler: CLDCompletionHandler? = nil) -> CLDFetchImageRequest {
-        let request = CLDFetchImageRequestImpl(url: url, networkCoordinator: networkCoordinator)
-        request.fetchImage()
-        request.responseImage(completionHandler)
-        request.progress(progress)
-        return request
-    }
 }
