+++ Cloudinary/Classes/ios/NetworkRequest/CLDFetchImageRequestImpl.swift
