+++ Cloudinary/Classes/ios/Uploader/CLDImagePreprocessChain.swift
