+++ Cloudinary/Classes/ios/Uploader/CLDPreprocessHelpers.swift
